Add Ubuntu scripts
This commit add Ubuntu installation and setup scripts and in order to
achieve that it was necessary major changes in the dotfiles
organization.

Function files are responsable to handle OS specific installation
details and preferences setup. They export only self-contained functions
in order to improve the testability of each step.

.shell/setup.sh is responsable to detect the user OS and orchestrate the
call order for each installation and setup functions.

To keep some consistency between OSX and Linux this PR add two aliases
for xclip when running in a Linux OS named pbcopy and pbpaste.

To easy recognize dotfile messages it was add a prefix (~>) in each
echo.