Fix integration with Linux clipboard of vim and tmux
tmux versions above or equal 2.4 require a custom configuration using
xclip to send visual selection to Linux clipboard.

The standard vim package for Ubuntu doesn't find * and + registers (both
responsable to send vim's buffer to OS clipboard). In order to fix this
issue we changed vim package for vim-gnome which has the integration out
of the box.

Assuming that python3 custom configuration did broken the gnome-terminal
this commit is also removing it from function-ubuntu until the problem
be diagnosticated.

Change vim rule markup to 80 and 120 chars due to typed languages, such
as TypeScript, requires more space.