Major rewrite of my Emacs and system management (#14)
The rewrite aims purpose, isolation, and testability. If I don't know why
something is here, it will be removed without thinking twice. I should be able
to delete a whole chunk of my configuration without the fear of break the
installation of something elsewhere. I have minimal confidence that installing
my dotfiles from scratch is working with a high level of automation.

It was done in two touchpoints: Emacs and system dependencies.

For the Emacs part, I rewrote my personal configuration from scratch. At the
moment, I have been focused on improving my developer experience with
Ruby. Therefore, some other language support are still missing (e.g. Clojure).

I switched `el-get' for `scratch' and `use-package' because of the isolation
the mindset of my dotfiles and because the community seems to be moving on that
direction.

I've introduced Org on my daily workflow for note-taking. Thus, there's a chunk
of my Emacs focused on it now.

Of course, I didn't write everything myself - I often get inspired by other
people dotfiles. Here's a list of my biggest references:

- https://github.com/ilmotta/dotfiles
- https://github.com/hrs/dotfiles
- https://github.com/hlissner/doom-emacs
- https://github.com/syl20bnr/spacemacs

For closing the Emacs section, I also switched from terminal to GUI. Hence, I've
introduced some configurations and packages only for it (see
`my-packages-gui.el').

The second part of the rewrite is how system dependencies are managed. For that
I've introduced Ansible. Everything done by it could be achieved with shell
scripts, nevertheless, I'd a personal curiosity on the tool and that was my main
driver to use it in my dotfiles. Although, isolation and testability of
dependencies and configurations are easily achievable with Ansible - which played
nice with my dotfiles.

Alongside that, I've been trying to improve my dotfiles documentation
aiming long term maintenance (I tend to forget things quickly). Thus, there's a
huge improvement in the README area and code comments everywhere.