Improve org-roam templates and bindings
Add the day of the week in the daily template title.

Add two psychological questions in the daily template to track mood change.

Add one question in the daily template to track important work to be done.

Improve work template by adding a headling equals to the file title.

Add bindings to easily navigate on org-roam today, tomorrow, yesterday and find
a date.